Django test marked as failing.
The Django test is currently known to fail with "GraphError" exceptions and has
now been marked as such. For organization, this test has been moved to a
separate script in the new "tests/functional/test_hooks" subdirectory.
